refactor folder structure and add datatype
