+++ src/app/features/skill/bulk-link-external-profiles/bulk-link-external-profiles.stories.ts
@@ -2,10 +2,10 @@ import { applicationConfig, Meta, StoryObj } from "@storybook/angular";
 import { TestEntity } from "../../../utils/test-utils/TestEntity";
 import { BulkLinkExternalProfilesComponent } from "./bulk-link-external-profiles.component";
 import { Entity } from "../../../core/entity/model/entity";
-import { SkillApiService } from "../skill-api.service";
+import { SkillApiService } from "../skill-api/skill-api.service";
 import { StorybookBaseModule } from "../../../utils/storybook-base.module";
 import { importProvidersFrom } from "@angular/core";
-import { mockSkillApi } from "../skill-api-mock";
+import { mockSkillApi } from "../skill-api/skill-api-mock";
 import { ExternalProfileLinkConfig } from "../external-profile-link-config";
 import { FormFieldConfig } from "../../../core/common-components/entity-form/FormConfig";
 import { throwError } from "rxjs";
